Duplicate misc folders confusion resolved
